Refactor runs.py to package
